Split comment edit tasks
