+++ handlers/blogs/blogsCommentEditCancelTask.go
@@ -1,0 +1,36 @@
+package blogs
+
+import (
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/gorilla/mux"
+
+	notif "github.com/arran4/goa4web/internal/notifications"
+	"github.com/arran4/goa4web/internal/tasks"
+)
+
+// CancelTask cancels comment editing.
+type CancelTask struct{ tasks.TaskString }
+
+var cancelTask = &CancelTask{TaskString: TaskCancel}
+
+var _ tasks.Task = (*CancelTask)(nil)
+var _ notif.AdminEmailTemplateProvider = (*CancelTask)(nil)
+
+func (CancelTask) AdminEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("adminNotificationBlogCommentCancelEmail")
+}
+
+func (CancelTask) AdminInternalNotificationTemplate() *string {
+	v := notif.NotificationTemplateFilenameGenerator("adminNotificationBlogCommentCancelEmail")
+	return &v
+}
+
+func (CancelTask) Action(w http.ResponseWriter, r *http.Request) any {
+	vars := mux.Vars(r)
+	blogId, _ := strconv.Atoi(vars["blog"])
+	http.Redirect(w, r, fmt.Sprintf("/blogs/blog/%d/comments", blogId), http.StatusTemporaryRedirect)
+	return nil
+}
