+++ handlers/blogs/blogsCommentEditReplyTask.go
@@ -4,13 +4,13 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"github.com/arran4/goa4web/core/consts"
 	"log"
 	"net/http"
 	"strconv"
 
 	"github.com/arran4/goa4web/core"
 	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
 	"github.com/arran4/goa4web/internal/db"
 	notif "github.com/arran4/goa4web/internal/notifications"
 	"github.com/arran4/goa4web/internal/tasks"
@@ -92,30 +92,6 @@ func (EditReplyTask) Action(w http.ResponseWriter, r *http.Request) any {
 		}
 	}
 
-	http.Redirect(w, r, fmt.Sprintf("/blogs/blog/%d/comments", blogId), http.StatusTemporaryRedirect)
-	return nil
-}
-
-// CancelTask cancels comment editing.
-type CancelTask struct{ tasks.TaskString }
-
-var cancelTask = &CancelTask{TaskString: TaskCancel}
-
-var _ tasks.Task = (*CancelTask)(nil)
-var _ notif.AdminEmailTemplateProvider = (*CancelTask)(nil)
-
-func (CancelTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationBlogCommentCancelEmail")
-}
-
-func (CancelTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationBlogCommentCancelEmail")
-	return &v
-}
-
-func (CancelTask) Action(w http.ResponseWriter, r *http.Request) any {
-	vars := mux.Vars(r)
-	blogId, _ := strconv.Atoi(vars["blog"])
 	http.Redirect(w, r, fmt.Sprintf("/blogs/blog/%d/comments", blogId), http.StatusTemporaryRedirect)
 	return nil
 }
